% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{hfin}
\alias{hfin}
\title{Shiny App for Health Finance}
\usage{
hfin()
}
\value{
shiny application object
}
\description{
Opens the shiny interface for the health finance functionality
provided by the \code{healthfinance} package. The interface currently
consists of 3 tabs: (1) import; (2) model; and (3) export.
}
